Module Level Implementation and Unit Testing
Implement the Lab Results and Consultation Feature in the UPSTAC API application. In order to implement this, follow the instructions given below:

With respect to the Tester, complete the code in 'LabRequestController' class to return the list of test requests assigned to current tester. Specifically, you need to implement the "getForTester()" method.
With respect to the Doctor, complete the code in ‘ConsultationController’ class to view all the test requests, the test requests assigned to current doctor, assign consultations to themselves and update the doctor suggestions. Specifically, you need to implement the
"getForConsultations()", "getForDoctor()", "assignForConsultation()" and the "updateConsultation()" methods.